+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.4.0 (in development)
 - Support the `%{c}h` log directive
 - `%R` can now be `None`
 - **Bugfix**: `%u` can now match the string `""` (two double quotes)
+- Support `mod_ssl`'s `%{*}c` and `%{*}x` directives
 
 v0.3.0 (2019-05-12)
 -------------------